requirements.txt
azure-functions
numpy
pandas
scikit-learn
tensorflow
joblib
